refactor(logger/config): 重命名日志代码文件名、解耦中间件、config统一
