+++ internal/middleware/request_id.go
@@ -1,0 +1,27 @@
+package middleware
+
+import (
+	"github.com/TogetherForStudy/jxust-yqlx-server/pkg/constant"
+	"github.com/TogetherForStudy/jxust-yqlx-server/pkg/logger"
+
+	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
+)
+
+func RequestID() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		requestID := c.Request.Header.Get("X-Request-ID")
+		if requestID == "" {
+			requestID = uuid.NewString()
+			c.Request.Header.Set("X-Request-ID", requestID)
+		}
+		c.Header("X-Request-ID", requestID)
+		c.Set(constant.RequestID, requestID)
+
+		ctx := logger.EnrichContext(c.Request.Context(), map[string]any{
+			"request_id": requestID,
+		})
+		c.Request = c.Request.WithContext(ctx)
+		c.Next()
+	}
+}
